Allow reading parquet data from a buffer
Although parquet was designed with for huge data sets that definitely don't fit in memory, it is also useful for compressed data tables of just a few MB that you want to load in memory in the browser.  Being able to use a buffer directly for this case is nice.